+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 [Full Changelog](https://github.com/maximeg/email_inquire/compare/v0.6.0...master)
 
+**Changed:**
+
+- Update one-time providers list to 3030 domains ([maximeg](https://github.com/maximeg)
+
 ## [v0.6.0](https://github.com/maximeg/email_inquire/tree/v0.6.0) (2017-04-16)
 [Full Changelog](https://github.com/maximeg/email_inquire/compare/v0.5.0...v0.6.0)